Added an rss-feed and rss-item element, plus update previous elements to overcome transform problems
The RSS Feed elements don't cover all the rss xml elements, but it covers enough to be useful.  The date, because of the non standard way this is implemented, is just kept as plain text. Future versions may involve custom parsers.

Polymer pubspec transformers will rewrite attributes like href and src.  This is not desired where these are identifier strings for icons.  Overcoming this by changing the attribute to something like _icon, _href or _src.